Fix some stuff for release (#164)
* Use released versions of tinybmp/tinytga

* Move README.md into embedded-graphics/

Because README.md is included in the code when released, it needs to be
in the same directory as the crate root.

This change moves the project's README.md into embedded-graphics/ and,
as an interim measure, symlinks the readme into the repository root so
something shows up when browsing on Github.

* Fix link to Dimensions

Not sure how this slipped through the other PR - it was running the same
tests!