Optimise Triangle.contains() performance
Uses a method found on StackOverflow to do a faster point-in-triangle
check. Because rendered triangles aren't mathematically accurate due to
some point centers falling outside the lines definining the triangle,
this change adds a final check over all points in the border lines to
ensure the contains() behaviour matches up with what is drawn using the
points() method.