Faster triangle point intersection (#390)
* Document some Triangle behaviour

* Add triangle contains point example

Useful for debugging this PR, but maybe not particularly useful outside
of that usage. Could easily be removed, or maybe we remove the
`primitives-triangles.rs` demo and replace it with this.

* Optimise Triangle.contains() performance

Uses a method found on StackOverflow to do a faster point-in-triangle
check. Because rendered triangles aren't mathematically accurate due to
some point centers falling outside the lines definining the triangle,
this change adds a final check over all points in the border lines to
ensure the contains() behaviour matches up with what is drawn using the
points() method.

* Add triangle contains point benchmark

* Colinear triangles can never contain points

Colinear triangles are still rendered, however. This maintains
consistency with other primitives and makes sense - the edges are still
there, just all on the same line.

* Remove contains example

* Improve test to draw inside mock display area