Authenticate CircleCI image pulls for custom image
This is due to Docker Hub's new rate limits for unauthenticated pulls -
by IP address. There are still limits for authenticated pulls, but are
limited now per Docker repo, not by IP address, so we don't hit limits
when pulling from CircleCI's IP addresses.

The `DOCKERHUB_PASSWORD` env var has already been added to the CircleCI
env var secrets thing. This uses my personal account, but it's not too
hard to change it to anything else in the future as (afaik) no special
permissions are required beyond login as the image is public.