Move to new cimg-based CircleCI images
Not sure what's better, but CircleCI probably wants us to move to this
new image type.

I also had to pin `just` to 0.6 as the new versions use `matches!()`
which aren't available in our MSRV of 1.40.0

This image tag is suffixed with `-cimg` as it uses the same Rust
version, but switches base repositories. Future images should just use
the Rust version without `-cimg` suffix as normal.