Authenticate CircleCI image pulls for custom image (#445)
* Authenticate CircleCI image pulls for custom image

This is due to Docker Hub's new rate limits for unauthenticated pulls -
by IP address. There are still limits for authenticated pulls, but are
limited now per Docker repo, not by IP address, so we don't hit limits
when pulling from CircleCI's IP addresses.

The `DOCKERHUB_PASSWORD` env var has already been added to the CircleCI
env var secrets thing. This uses my personal account, but it's not too
hard to change it to anything else in the future as (afaik) no special
permissions are required beyond login as the image is public.

* Move to new cimg-based CircleCI images

Not sure what's better, but CircleCI probably wants us to move to this
new image type.

I also had to pin `just` to 0.6 as the new versions use `matches!()`
which aren't available in our MSRV of 1.40.0

This image tag is suffixed with `-cimg` as it uses the same Rust
version, but switches base repositories. Future images should just use
the Rust version without `-cimg` suffix as normal.