Detect nearly-parallel numerical inaccuracy
Not too happy with this solution, but it fixes the original problem so
at least there's that.

I don't understand the maths behind it at _all_, this just seemed to
work for the cases in #471. Hopefully no other bugs are introduced by
this change.

I'm very open to a better solution, or at least an explanation of why my
fix works.

Closes #471.