Fix invalid MonoStyle text dimensions calculation
The method len() on &str does return the number of bytes of a character,
not the number of characters.
Using characters such as '°' (which are multiple bytes long) caused an
incorrect calculation of the text bounding box.

This commit fixes this by calling .chars().count() on &str.
Also, a test was added to prevent false dimension calculations of
multi-byte character strings in the future.